updated hooks for play server

+++ redformgithub/hook.php
@@ -10,10 +10,33 @@ $convert = rawurldecode($str);
 
 parse_str($convert);
 
-$targetBranch = 'maersk-main';
+parse_str($convert);
+$payload = json_decode($convert);
+
+if (strstr($_SERVER['SERVER_NAME'], 'play'))
+{
+	$targetBranch = 'maersk-version-qa';
+	$basepath = '/home/play';
+}
+else
+{
+	$targetBranch = 'maersk-main';
+	$basepath = '/home/staging';
+}
+
+if (!strstr($payload->ref, $targetBranch))
+{
+	echo 'Another branch was updated';
+
+	return true;
+}
+else
+{
+	echo $targetBranch . ' was updated';
+}
 
 // Update repo
-$cmd = 'cd /home/staging/git/redFORM2.5; git fetch --all 2<&1; ';
+$cmd = 'cd ' . $basepath . '/git/redFORM2.5; git fetch --all 2<&1; ';
 $cmd .= 'git reset --hard origin/' . $targetBranch . ' 2<&1; ';
 $cmd .= 'git submodule update 2<&1; ';
 
@@ -21,7 +44,7 @@ $cmd .= 'git submodule update 2<&1; ';
 $cmd .= 'phing 2<&1; ';
 
 // Update db
-$cmd .= 'php /home/staging/public_html/redformgithub/redInstall.php --extension=redform; ';
+$cmd .= 'php ' . $basepath . '/public_html/redformgithub/redInstall.php --extension=redform; ';
 
 $output = shell_exec($cmd);
